+++ internal/meshes/istio/pipeline/authorization_policy.go
@@ -3,7 +3,7 @@ package pipeline
 import (
 	"log"
 
-	"github.com/layer5io/meshsync/internal/model"
+	"github.com/layer5io/meshsync/pkg/model"
 	broker "github.com/layer5io/meshsync/pkg/broker"
 	discovery "github.com/layer5io/meshsync/pkg/discovery"
 	"github.com/myntra/pipeline"
